% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a ggplot box plot}
\usage{
gbox(
  values,
  categories = NULL,
  color = "black",
  fill = "lightblue",
  title = "Box Plot",
  value_lab = "Values",
  cat_lab = "Categories",
  horizontal = FALSE
)
}
\arguments{
\item{values}{A numeric vector of data values.}

\item{categories}{(Optional) A factor vector specifying categories for grouping the values.}

\item{color}{The color of the box plot outlines (default is "black").}

\item{fill}{The color to fill the box plots (default is "lightblue").}

\item{title}{The title of the plot (default is "Box Plot").}

\item{value_lab}{The label for the y-axis (default is "Values").}

\item{cat_lab}{The label for the x-axis (default is "Categories").}

\item{horizontal}{Logical; if TRUE, the box plots are horizontal; if FALSE (default), they are vertical.}
}
\value{
A ggplot object representing the box plot.
}
\description{
This function generates a ggplot box plot for a given set of values and optional categories.
}
\examples{
# Box plot without categories
gbox(c(1, 2, 3, 4, 5))

# Box plot with categories
gbox(c(1, 2, 3, 4, 5), categories = c("A", "A", "B", "B", "B"))

}
