% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{Create a ggplot histogram}
\usage{
ghist(
  x,
  point_color = "black",
  line_color = "black",
  fill_color = "lightblue",
  title = "Histogram",
  xlab = "X",
  ylab = "Frequency",
  bins = 0
)
}
\arguments{
\item{x}{A numeric vector for which the histogram will be created.}

\item{point_color}{The color of the points (default is "black").}

\item{line_color}{The color of the lines (default is "black").}

\item{fill_color}{The color of the fill (default is "lightblue").}

\item{title}{The title of the plot (default is "Histogram").}

\item{xlab}{The x-axis label (default is "X").}

\item{ylab}{The y-axis label (default is "Frequency").}

\item{bins}{The number of bins in the histogram (default is 0, which auto-computes bins).}
}
\value{
A ggplot histogram object.
}
\description{
This function creates a ggplot histogram of a numeric vector.
}
\examples{
data(mtcars)
ghist(mtcars$mpg, bins = 0)
ghist(mtcars$mpg, bins = 3)

}
